+++ lib/modules/storage/storage_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: storage
-// Updated at: 2026-07-03 22:07:45
-// Commit: feat(storage): add bi-directional event bus handler for storage
+// Updated at: 2026-07-05 11:17:41
+// Commit: feat(storage): implement shimmer loading skeleton in storage
 
 class storageComponentState {
-  final String commitRef = '2558';
-  final String lastMessage = 'feat(storage): add bi-directional event bus handler for storage';
-  final DateTime timestamp = DateTime.parse('2026-07-03T22:07:45+05:00');
+  final String commitRef = '2594';
+  final String lastMessage = 'feat(storage): implement shimmer loading skeleton in storage';
+  final DateTime timestamp = DateTime.parse('2026-07-05T11:17:41+05:00');
 }